refactor(manipulation): split RoboPlan planner from world
