+++ dimos/manipulation/planning/factory.py
@@ -27,8 +27,8 @@ from dimos.manipulation.planning.kinematics.config import (
     kinematics_config_from_name,
 )
 from dimos.manipulation.planning.planners.config import ManipulationPlannerConfig
+from dimos.manipulation.planning.planners.roboplan_config import RoboPlanPlannerConfig
 from dimos.manipulation.planning.spec.protocols import PlannerSpec
-from dimos.manipulation.planning.world.roboplan_config import RoboPlanPlannerConfig
 from dimos.manipulation.visualization.config import (
     ManipulationVisualizationConfig,
     NoManipulationVisualizationConfig,
@@ -148,8 +148,8 @@ def create_planner(
 ) -> PlannerSpec:
     """Create a motion planner from its typed backend configuration.
 
-    RoboPlan-native planning is scene/backend-coupled, so its config returns the
-    RoboPlan world object itself as the planner.
+    RoboPlan-native planning is scene/backend-coupled, so its planner is bound to
+    the supplied RoboPlan world.
     """
     if config.backend == "rrt_connect":
         from dimos.manipulation.planning.planners.rrt_planner import RRTConnectPlanner
@@ -158,12 +158,9 @@ def create_planner(
     if config.backend == "roboplan":
         if world_backend != "roboplan" or world is None:
             raise ValueError(_ROBOPLAN_PLANNER_REQUIRES_ROBOPLAN_WORLD)
-        from dimos.manipulation.planning.world.roboplan_world import RoboPlanWorld
+        from dimos.manipulation.planning.planners.roboplan_planner import RoboPlanPlanner
 
-        if not isinstance(world, RoboPlanWorld):
-            raise ValueError("RoboPlan-native planner requires a RoboPlan world planner object")
-        world.configure_planner(config)
-        return world
+        return RoboPlanPlanner(world, config)
 
     raise TypeError(f"Unsupported planner config: {type(config).__name__}")
