+++ dimos/manipulation/test_roboplan.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Pure-Python tests for the optional RoboPlan world adapter."""
+"""Pure-Python tests for the optional RoboPlan world and planner adapters."""
 
 from __future__ import annotations
 
@@ -35,15 +35,15 @@ from dimos.manipulation.planning.groups.models import (
     PlanningGroupSelection,
 )
 from dimos.manipulation.planning.groups.registry import PlanningGroupRegistry
-from dimos.manipulation.planning.planners.rrt_planner import RRTConnectPlanner
-from dimos.manipulation.planning.spec.config import RobotModelConfig
-from dimos.manipulation.planning.spec.enums import ObstacleType, PlanningStatus
-from dimos.manipulation.planning.spec.models import Obstacle
-from dimos.manipulation.planning.world.roboplan_config import (
+from dimos.manipulation.planning.planners.roboplan_config import (
     RoboPlanCartesianPathConfig,
     RoboPlanPathShortcuttingConfig,
     RoboPlanPlannerConfig,
 )
+from dimos.manipulation.planning.planners.rrt_planner import RRTConnectPlanner
+from dimos.manipulation.planning.spec.config import RobotModelConfig
+from dimos.manipulation.planning.spec.enums import ObstacleType, PlanningStatus
+from dimos.manipulation.planning.spec.models import Obstacle
 from dimos.msgs.geometry_msgs.PoseStamped import PoseStamped
 from dimos.msgs.geometry_msgs.Quaternion import Quaternion
 from dimos.msgs.geometry_msgs.Transform import Transform
@@ -490,8 +490,6 @@ def _make_world(
     module = _import_roboplan_world(fake_roboplan)
 
     world = module.RoboPlanWorld()
-    if planner_config is not None:
-        world.configure_planner(planner_config)
     robot_id = world.add_robot(robot_config)
     world.finalize()
     world.sync_from_joint_state(
@@ -500,6 +498,11 @@ def _make_world(
             name=list(robot_config.joint_names),
             position=[0.0] * len(robot_config.joint_names),
         ),
+    )
+    planner_module = _import_roboplan_planner(fake_roboplan)
+    _PLANNERS_BY_WORLD[world] = planner_module.RoboPlanPlanner(
+        world,
+        planner_config or RoboPlanPlannerConfig(),
     )
     return world, robot_id
 
@@ -564,12 +567,40 @@ def _import_roboplan_world(fake_roboplan: None) -> ModuleType:
     return importlib.import_module(module_name)
 
 
-def test_roboplan_bindings_are_imported_at_module_load(fake_roboplan: None) -> None:
-    module = _import_roboplan_world(fake_roboplan)
+def _import_roboplan_planner(fake_roboplan: None) -> ModuleType:
+    module_name = "dimos.manipulation.planning.planners.roboplan_planner"
+    if module_name in sys.modules:
+        return importlib.reload(sys.modules[module_name])
+    return importlib.import_module(module_name)
 
-    assert module.roboplan_core.Scene is FakeScene
-    assert module.roboplan_rrt.RRT is FakeRRT
-    assert module.roboplan_cartesian.CartesianPathPlanner is FakeCartesianPathPlanner
+
+_PLANNERS_BY_WORLD: dict[Any, Any] = {}
+
+
+def _planner_for(world: Any) -> Any:
+    if world not in _PLANNERS_BY_WORLD:
+        planner_module = _import_roboplan_planner(None)
+        _PLANNERS_BY_WORLD[world] = planner_module.RoboPlanPlanner(
+            world,
+            RoboPlanPlannerConfig(),
+        )
+    return _PLANNERS_BY_WORLD[world]
+
+
+def test_roboplan_bindings_are_imported_at_module_load(fake_roboplan: None) -> None:
+    world_module = _import_roboplan_world(fake_roboplan)
+    planner_module = _import_roboplan_planner(fake_roboplan)
+
+    assert world_module.roboplan_core.Scene is FakeScene
+    assert planner_module.roboplan_rrt.RRT is FakeRRT
+    assert planner_module.roboplan_cartesian.CartesianPathPlanner is FakeCartesianPathPlanner
+
+
+def test_roboplan_planner_rejects_non_roboplan_world(fake_roboplan: None) -> None:
+    planner_module = _import_roboplan_planner(fake_roboplan)
+
+    with pytest.raises(TypeError, match="requires a RoboPlanWorld"):
+        planner_module.RoboPlanPlanner(object(), RoboPlanPlannerConfig())
 
 
 def test_robot_registration_finalization_and_joint_limits(
@@ -1323,7 +1354,7 @@ def test_native_planner_converts_path(fake_roboplan: None, robot_config: RobotMo
 
     start = JointState(name=["joint1", "joint2"], position=[0.0, 0.0])
     goal = JointState(name=["joint1", "joint2"], position=[0.4, 0.2])
-    result = world.plan_joint_path(world, robot_id, start, goal, timeout=1.0)
+    result = _planner_for(world).plan_joint_path(world, robot_id, start, goal, timeout=1.0)
 
     assert result.status == PlanningStatus.SUCCESS
     assert [state.position for state in result.path] == [[0.0, 0.0], [0.2, 0.1], [0.4, 0.2]]
@@ -1358,7 +1389,7 @@ def test_native_planner_shortcuts_path_with_configured_options(
         side_effect=endpoints_only,
     )
 
-    result = world.plan_joint_path(
+    result = _planner_for(world).plan_joint_path(
         world,
         robot_id,
         JointState(name=["joint1", "joint2"], position=[0.0, 0.0]),
@@ -1389,7 +1420,7 @@ def test_native_planner_can_disable_path_shortcutting(
 
     shortcut = mocker.patch.object(FakePathShortcutter, "shortcut", autospec=True)
 
-    result = world.plan_joint_path(
+    result = _planner_for(world).plan_joint_path(
         world,
         robot_id,
         JointState(name=["joint1", "joint2"], position=[0.0, 0.0]),
@@ -1414,7 +1445,7 @@ def test_native_planner_uses_raw_path_when_shortcutting_fails(
         side_effect=RuntimeError("shortcut failed"),
     )
 
-    result = world.plan_joint_path(
+    result = _planner_for(world).plan_joint_path(
         world,
         robot_id,
         JointState(name=["joint1", "joint2"], position=[0.0, 0.0]),
@@ -1440,7 +1471,7 @@ def test_native_planner_surfaces_unexpected_shortcutting_error(
     )
 
     with pytest.raises(TypeError, match="unexpected shortcut integration error"):
-        world.plan_joint_path(
+        _planner_for(world).plan_joint_path(
             world,
             robot_id,
             JointState(name=["joint1", "joint2"], position=[0.0, 0.0]),
@@ -1470,7 +1501,7 @@ def test_native_planner_uses_raw_path_when_shortcutting_changes_endpoint(
         side_effect=changed_endpoint,
     )
 
-    result = world.plan_joint_path(
+    result = _planner_for(world).plan_joint_path(
         world,
         robot_id,
         JointState(name=["joint1", "joint2"], position=[0.0, 0.0]),
@@ -1495,7 +1526,7 @@ def test_native_planner_uses_raw_path_when_shortcutting_returns_empty_path(
         return_value=FakeJointPath(["joint1", "joint2"], []),
     )
 
-    result = world.plan_joint_path(
+    result = _planner_for(world).plan_joint_path(
         world,
         robot_id,
         JointState(name=["joint1", "joint2"], position=[0.0, 0.0]),
@@ -1507,14 +1538,28 @@ def test_native_planner_uses_raw_path_when_shortcutting_returns_empty_path(
     assert [state.position for state in result.path] == [[0.0, 0.0], [0.2, 0.1], [0.4, 0.2]]
 
 
-def test_roboplan_planner_configuration_is_fixed_at_finalization(
+def test_roboplan_planner_is_distinct_and_bound_to_world(
     fake_roboplan: None,
     robot_config: RobotModelConfig,
 ) -> None:
     world, _ = _make_world(fake_roboplan, robot_config)
+    planner = _planner_for(world)
 
-    with pytest.raises(RuntimeError, match="after world finalization"):
-        world.configure_planner(RoboPlanPlannerConfig())
+    assert planner is not world
+    assert planner._world is world
+    assert not hasattr(world, "plan_joint_path")
+
+
+def test_roboplan_planner_copies_configuration(
+    fake_roboplan: None,
+    robot_config: RobotModelConfig,
+) -> None:
+    config = RoboPlanPlannerConfig(path_shortcutting=RoboPlanPathShortcuttingConfig(enabled=False))
+    world, _ = _make_world(fake_roboplan, robot_config, config)
+
+    config.path_shortcutting.enabled = True
+
+    assert not _planner_for(world)._config.path_shortcutting.enabled
 
 
 def test_native_planning_blocks_obstacle_replacement(
@@ -1548,7 +1593,9 @@ def test_native_planning_blocks_obstacle_replacement(
     start = JointState(name=["joint1", "joint2"], position=[0.0, 0.0])
     goal = JointState(name=["joint1", "joint2"], position=[0.2, 0.1])
     planning_thread = threading.Thread(
-        target=lambda: world.plan_joint_path(world, robot_id, start, goal, timeout=1.0)
+        target=lambda: _planner_for(world).plan_joint_path(
+            world, robot_id, start, goal, timeout=1.0
+        )
     )
     planning_thread.start()
     assert planning_started.wait(1.0)
@@ -1573,7 +1620,7 @@ def test_native_planner_names_path_from_robot_config_when_start_is_unnamed(
 
     start = JointState(name=[], position=[0.0, 0.0])
     goal = JointState(name=["joint1", "joint2"], position=[0.4, 0.2])
-    result = world.plan_joint_path(world, robot_id, start, goal, timeout=1.0)
+    result = _planner_for(world).plan_joint_path(world, robot_id, start, goal, timeout=1.0)
 
     assert result.status == PlanningStatus.SUCCESS
     assert [state.name for state in result.path] == [["joint1", "joint2"]] * 3
@@ -1585,7 +1632,7 @@ def test_native_selected_planner_returns_global_selected_joint_names(
     world, _ = _make_world(fake_roboplan, robot_config)
     selection = _selection((robot_config,), "arm/manipulator")
 
-    result = world.plan_selected_joint_path(
+    result = _planner_for(world).plan_selected_joint_path(
         world,
         selection,
         JointState(name=["arm/joint1", "arm/joint2"], position=[0.0, 0.0]),
@@ -1604,7 +1651,7 @@ def test_native_selected_planner_accepts_local_joint_names(
     world, _ = _make_world(fake_roboplan, robot_config)
     selection = _selection((robot_config,), "arm/manipulator")
 
-    result = world.plan_selected_joint_path(
+    result = _planner_for(world).plan_selected_joint_path(
         world,
         selection,
         JointState(name=["joint2", "joint1"], position=[0.0, 0.0]),
@@ -1630,7 +1677,7 @@ def test_native_selected_planner_rejects_multi_group_selection(
     world, _ = _make_world(fake_roboplan, config)
     selection = _selection((config,), "arm/left", "arm/right")
 
-    result = world.plan_selected_joint_path(
+    result = _planner_for(world).plan_selected_joint_path(
         world,
         selection,
         JointState(name=list(selection.joint_names), position=[0.0, 0.0]),
@@ -1651,7 +1698,7 @@ def test_native_planner_coordinates_groups_across_two_robots(
         "arm/manipulator",
     )
 
-    result = world.plan_selected_joint_path(
+    result = _planner_for(world).plan_selected_joint_path(
         world,
         selection,
         JointState(name=list(selection.joint_names), position=[0.0, 0.0, 0.0, 0.0]),
@@ -1695,7 +1742,7 @@ def test_cartesian_planner_returns_timed_global_joint_states_and_options(
         "max_attempts_per_step": 8,
     }
 
-    result = world.plan_cartesian_path(
+    result = _planner_for(world).plan_cartesian_path(
         world,
         selection,
         JointState(name=list(selection.joint_names), position=[0.0, 0.0]),
@@ -1744,7 +1791,7 @@ def test_cartesian_zero_rotation_preserves_start_orientation(
     world, _ = _make_world(fake_roboplan, robot_config)
     selection = _selection((robot_config,), "arm/manipulator")
 
-    result = world.plan_cartesian_path(
+    result = _planner_for(world).plan_cartesian_path(
         world,
         selection,
         JointState(name=list(selection.joint_names), position=[0.0, 0.0]),
@@ -1774,7 +1821,7 @@ def test_cartesian_supports_mixed_targets_and_shared_multi_group_timing(
         "arm/manipulator",
     )
 
-    result = world.plan_cartesian_path(
+    result = _planner_for(world).plan_cartesian_path(
         world,
         selection,
         JointState(name=list(selection.joint_names), position=[0.0] * 4),
@@ -1810,7 +1857,7 @@ def test_cartesian_allows_auxiliary_groups(
         "right/manipulator",
     )
 
-    result = world.plan_cartesian_path(
+    result = _planner_for(world).plan_cartesian_path(
         world,
         selection,
         JointState(name=list(selection.joint_names), position=[0.0] * 4),
@@ -1870,7 +1917,7 @@ def test_cartesian_rejects_invalid_requests(
     world, _ = _make_world(fake_roboplan, robot_config)
     selection = _selection((robot_config,), "arm/manipulator")
 
-    result = world.plan_cartesian_path(
+    result = _planner_for(world).plan_cartesian_path(
         world,
         selection,
         JointState(name=list(selection.joint_names), position=[0.0, 0.0]),
@@ -1890,7 +1937,7 @@ def test_cartesian_rejects_start_that_differs_from_scene(
     world, _ = _make_world(fake_roboplan, robot_config)
     selection = _selection((robot_config,), "arm/manipulator")
 
-    result = world.plan_cartesian_path(
+    result = _planner_for(world).plan_cartesian_path(
         world,
         selection,
         JointState(name=list(selection.joint_names), position=[0.1, 0.0]),
@@ -1916,7 +1963,7 @@ def test_cartesian_rejects_official_planner_failure(
         side_effect=ValueError("tracking failed"),
     )
 
-    result = world.plan_cartesian_path(
+    result = _planner_for(world).plan_cartesian_path(
         world,
         selection,
         JointState(name=list(selection.joint_names), position=[0.0, 0.0]),
@@ -1951,7 +1998,7 @@ def test_cartesian_postvalidation_checks_combined_multi_robot_state(
         side_effect=collides_only_when_both_arms_advance,
     )
 
-    result = world.plan_cartesian_path(
+    result = _planner_for(world).plan_cartesian_path(
         world,
         selection,
         JointState(name=list(selection.joint_names), position=[0.0] * 4),
@@ -2007,7 +2054,7 @@ def test_cartesian_postvalidation_checks_between_waypoints(
         side_effect=collides_only_mid_edge,
     )
 
-    result = world.plan_cartesian_path(
+    result = _planner_for(world).plan_cartesian_path(
         world,
         selection,
         JointState(name=list(selection.joint_names), position=[0.0, 0.0]),
@@ -2050,7 +2097,7 @@ def test_native_planner_preserves_other_robot_and_auxiliary_joint_state(
 
     mocker.patch.object(FakeRRT, "plan", autospec=True, side_effect=capture_scene_state)
 
-    result = world.plan_selected_joint_path(
+    result = _planner_for(world).plan_selected_joint_path(
         world,
         selection,
         JointState(name=list(selection.joint_names), position=[0.0, 0.0]),
@@ -2079,7 +2126,7 @@ def test_native_planner_waits_for_every_robot_state(
     )
     selection = _selection((robot_config, second_config), "arm/manipulator")
 
-    result = world.plan_selected_joint_path(
+    result = _planner_for(world).plan_selected_joint_path(
         world,
         selection,
         JointState(name=list(selection.joint_names), position=[0.0, 0.0]),
@@ -2104,7 +2151,7 @@ def test_native_planner_rejects_empty_path(
 
     start = JointState(name=["joint1", "joint2"], position=[0.0, 0.0])
     goal = JointState(name=["joint1", "joint2"], position=[0.4, 0.2])
-    result = world.plan_joint_path(world, robot_id, start, goal, timeout=1.0)
+    result = _planner_for(world).plan_joint_path(world, robot_id, start, goal, timeout=1.0)
 
     assert result.status == PlanningStatus.NO_SOLUTION
     assert result.path == []
