+++ dimos/manipulation/visualization/test_operator.py
@@ -19,6 +19,7 @@ from pathlib import Path
 from dimos.agents.skill_result import SkillResult
 from dimos.manipulation.planning.groups.models import PlanningGroup, PlanningGroupDefinition
 from dimos.manipulation.planning.groups.registry import PlanningGroupRegistry
+from dimos.manipulation.planning.planners.roboplan_config import RoboPlanCartesianPathConfig
 from dimos.manipulation.planning.spec.config import RobotModelConfig
 from dimos.manipulation.planning.spec.enums import IKStatus, PlanningStatus
 from dimos.manipulation.planning.spec.models import (
@@ -27,7 +28,6 @@ from dimos.manipulation.planning.spec.models import (
     PlanningGroupID,
     RobotName,
 )
-from dimos.manipulation.planning.world.roboplan_config import RoboPlanCartesianPathConfig
 from dimos.manipulation.visualization.operator import (
     CartesianTargetRequest,
     JointTargetRequest,
