+++ docs/capabilities/manipulation/index.md
@@ -127,7 +127,7 @@ dimos run xarm7-planner-coordinator \
 ```
 
 Existing RoboPlan deployments may therefore receive paths with fewer waypoints.
-Shortcutting configuration is fixed when the planning world is finalized.
+Shortcutting configuration is copied when the RoboPlan planner is constructed.
 
 The remaining options mirror RoboPlan's native path shortcutter:
 `seed`, `max_convergence_iters`, and `redundant_removal_iters`. If shortcutting
@@ -136,7 +136,7 @@ fails, planning returns the valid raw RRT path and logs a warning.
 RoboPlan Cartesian options are supplied per planning request:
 
 ```python skip
-from dimos.manipulation.planning.world.roboplan_config import (
+from dimos.manipulation.planning.planners.roboplan_config import (
     RoboPlanCartesianPathConfig,
 )
 
@@ -395,4 +395,6 @@ planner is locked for its whole native call.
 | [`robot/manipulators/xarm/blueprints/perception.py`](/dimos/robot/manipulators/xarm/blueprints/perception.py) | XArm perception blueprint |
 | [`teleop/keyboard/keyboard_teleop_module.py`](/dimos/teleop/keyboard/keyboard_teleop_module.py) | Keyboard teleop module |
 | [`planning/world/drake_world.py`](/dimos/manipulation/planning/world/drake_world.py) | Drake physics backend |
+| [`planning/world/roboplan_world.py`](/dimos/manipulation/planning/world/roboplan_world.py) | RoboPlan scene, state, and collision backend |
+| [`planning/planners/roboplan_planner.py`](/dimos/manipulation/planning/planners/roboplan_planner.py) | RoboPlan-native joint and Cartesian planner |
 | [`planning/planners/rrt_planner.py`](/dimos/manipulation/planning/planners/rrt_planner.py) | RRT-Connect motion planner |
